coverage percents

+++ plugins/org.eclipse.osee.coverage.test/src/org/eclipse/osee/coverage/test/CoveragePackageImportTest.java
@@ -93,7 +93,7 @@ public class CoveragePackageImportTest {
 
       System.out.println(CoverageUtil.printTree(coverageImport));
 
-      Assert.assertEquals(49, coverageImport.getCoveragePercent());
+      Assert.assertEquals(49, coverageImport.getCoveragePercent().intValue());
       Assert.assertEquals(0, coverageImport.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coverageImport.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(60, coverageImport.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -205,7 +205,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(60, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(122, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(49, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(49, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(60, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -234,7 +234,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(60, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(122, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(49, loadedCp.getCoveragePercent());
+      Assert.assertEquals(49, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(60, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -252,7 +252,7 @@ public class CoveragePackageImportTest {
       // Check import results
       Assert.assertEquals(60, coverageImport.getCoverageItemsCovered().size());
       Assert.assertEquals(122, coverageImport.getCoverageItems().size());
-      Assert.assertEquals(49, coverageImport.getCoveragePercent());
+      Assert.assertEquals(49, coverageImport.getCoveragePercent().intValue());
       Assert.assertEquals(0, coverageImport.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coverageImport.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(60, coverageImport.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -315,7 +315,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(66, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(133, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(49, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(49, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(66, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -343,7 +343,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(66, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(133, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(49, loadedCp.getCoveragePercent());
+      Assert.assertEquals(49, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(66, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -416,7 +416,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(67, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(134, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(50, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(50, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(67, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -452,7 +452,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(67, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(134, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(50, loadedCp.getCoveragePercent());
+      Assert.assertEquals(50, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(67, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -532,7 +532,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(68, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(135, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(50, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(50, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(68, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -568,7 +568,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(68, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(135, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(50, loadedCp.getCoveragePercent());
+      Assert.assertEquals(50, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(68, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -655,7 +655,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(69, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(137, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(50, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(50, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(69, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -683,7 +683,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(69, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(137, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(50, loadedCp.getCoveragePercent());
+      Assert.assertEquals(50, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(69, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -783,7 +783,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(68, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(135, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(50, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(50, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(68, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -811,7 +811,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(68, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(135, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(50, loadedCp.getCoveragePercent());
+      Assert.assertEquals(50, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(68, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -935,7 +935,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(69, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(135, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(51, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(51, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(69, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -972,7 +972,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(69, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(135, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(51, loadedCp.getCoveragePercent());
+      Assert.assertEquals(51, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(69, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -1085,7 +1085,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(68, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(134, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(50, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(50, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(68, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -1123,7 +1123,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(68, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(134, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(50, loadedCp.getCoveragePercent());
+      Assert.assertEquals(50, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(68, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -1206,7 +1206,7 @@ public class CoveragePackageImportTest {
       // CoveragePackage should now have imported results
       Assert.assertEquals(70, coveragePackage.getCoverageItemsCovered().size());
       Assert.assertEquals(134, coveragePackage.getCoverageItems().size());
-      Assert.assertEquals(52, coveragePackage.getCoveragePercent());
+      Assert.assertEquals(52, coveragePackage.getCoveragePercent().intValue());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(70, coveragePackage.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
@@ -1234,7 +1234,7 @@ public class CoveragePackageImportTest {
 
       Assert.assertEquals(70, loadedCp.getCoverageItemsCovered().size());
       Assert.assertEquals(134, loadedCp.getCoverageItems().size());
-      Assert.assertEquals(52, loadedCp.getCoveragePercent());
+      Assert.assertEquals(52, loadedCp.getCoveragePercent().intValue());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Deactivated_Code).size());
       Assert.assertEquals(0, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Exception_Handling).size());
       Assert.assertEquals(70, loadedCp.getCoverageItemsCovered(CoverageOptionManager.Test_Unit).size());
