+++ plugins/org.eclipse.osee.coverage/src/org/eclipse/osee/coverage/model/CoverageTestUnit.java
@@ -87,8 +87,8 @@ public class CoverageTestUnit implements ICoverage {
    }
 
    @Override
-   public int getCoveragePercent() {
-      return 0;
+   public Double getCoveragePercent() {
+      return 0.0;
    }
 
    @Override
