+++ plugins/org.eclipse.osee.coverage/src/org/eclipse/osee/coverage/model/MessageCoverageItem.java
@@ -44,8 +44,8 @@ public class MessageCoverageItem implements ICoverage {
    }
 
    @Override
-   public int getCoveragePercent() {
-      return 0;
+   public Double getCoveragePercent() {
+      return 0.0;
    }
 
    @Override
